# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Password Strength Checker API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface PasswordInput {
  /** The password to analyze */
  password: string;
}

/**
 * Human-readable strength label
 */
export type PasswordAnalysisStrengthLabel = typeof PasswordAnalysisStrengthLabel[keyof typeof PasswordAnalysisStrengthLabel];


export const PasswordAnalysisStrengthLabel = {
  Very_Weak: 'Very Weak',
  Weak: 'Weak',
  Fair: 'Fair',
  Good: 'Good',
  Strong: 'Strong',
  Very_Strong: 'Very Strong',
} as const;

export interface ScoreBreakdown {
  lengthScore: number;
  varietyScore: number;
  patternScore: number;
  entropyScore: number;
}

export interface PasswordAnalysis {
  /** Numerical score 0-100 */
  score: number;
  /** Human-readable strength label */
  strengthLabel: PasswordAnalysisStrengthLabel;
  /** Strength level 0-5 */
  strengthLevel: number;
  /** Estimated entropy in bits */
  entropy: number;
  charCount: number;
  uppercaseCount: number;
  lowercaseCount: number;
  digitCount: number;
  symbolCount: number;
  /** Human-readable estimated crack time */
  crackTime: string;
  /** Improvement suggestions */
  suggestions: string[];
  breakdown: ScoreBreakdown;
  hasCommonPattern: boolean;
  hasRepeatedChars: boolean;
  hasSequentialChars: boolean;
  isDictionaryWord: boolean;
}

export interface GeneratorOptions {
  /**
     * @minimum 4
     * @maximum 128
     */
  length: number;
  useUppercase: boolean;
  useLowercase: boolean;
  useDigits: boolean;
  useSymbols: boolean;
  /** Exclude visually similar characters like 0/O, 1/l/I */
  excludeSimilar: boolean;
}

export interface GeneratedPassword {
  password: string;
  analysis: PasswordAnalysis;
}