# lib/api-zod/src/generated/types/generatedPassword.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Password Strength Checker API
 * OpenAPI spec version: 0.1.0
 */
import type { PasswordAnalysis } from './passwordAnalysis';

export interface GeneratedPassword {
  password: string;
  analysis: PasswordAnalysis;
}