# lib/api-zod/src/generated/types/generatorOptions.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Password Strength Checker API
 * OpenAPI spec version: 0.1.0
 */

export interface GeneratorOptions {
  /**
     * @minimum 4
     * @maximum 128
     */
  length: number;
  useUppercase: boolean;
  useLowercase: boolean;
  useDigits: boolean;
  useSymbols: boolean;
  /** Exclude visually similar characters like 0/O, 1/l/I */
  excludeSimilar: boolean;
}