# lib/api-zod/src/generated/types/passwordAnalysis.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Password Strength Checker API
 * OpenAPI spec version: 0.1.0
 */
import type { PasswordAnalysisStrengthLabel } from './passwordAnalysisStrengthLabel';
import type { ScoreBreakdown } from './scoreBreakdown';

export interface PasswordAnalysis {
  /** Numerical score 0-100 */
  score: number;
  /** Human-readable strength label */
  strengthLabel: PasswordAnalysisStrengthLabel;
  /** Strength level 0-5 */
  strengthLevel: number;
  /** Estimated entropy in bits */
  entropy: number;
  charCount: number;
  uppercaseCount: number;
  lowercaseCount: number;
  digitCount: number;
  symbolCount: number;
  /** Human-readable estimated crack time */
  crackTime: string;
  /** Improvement suggestions */
  suggestions: string[];
  breakdown: ScoreBreakdown;
  hasCommonPattern: boolean;
  hasRepeatedChars: boolean;
  hasSequentialChars: boolean;
  isDictionaryWord: boolean;
}