# lib/api-zod/src/generated/types/passwordInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Password Strength Checker API
 * OpenAPI spec version: 0.1.0
 */

export interface PasswordInput {
  /** The password to analyze */
  password: string;
}