# lib/api-zod/src/generated/types/scoreBreakdown.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Password Strength Checker API
 * OpenAPI spec version: 0.1.0
 */

export interface ScoreBreakdown {
  lengthScore: number;
  varietyScore: number;
  patternScore: number;
  entropyScore: number;
}